# webapi/Controllers/QuotesController.cs
using Microsoft.AspNetCore.Mvc;
using Quotes.Storage.Interface;

namespace webapi.Controllers;

[ApiController]
[Route("api/[controller]")]
public class QuotesController : ControllerBase
{
    private readonly ILogger<QuotesController> _logger;
    private readonly global::Quotes.Quotes quotes;

    public QuotesController(ILogger<QuotesController> logger, IQuoteStorage storage)
    {
        _logger = logger;
        this.quotes = new global::Quotes.Quotes(storage);
    }

    [HttpGet]
    [Route("")]
    public async Task GetQuotes([FromQuery] string? movie, [FromQuery] string? actor, [FromQuery] int? start, [FromQuery] int? count)
    {
        // todo: pagination

        Console.WriteLine("doo");
        // Todo: rethink return objects
        await this.quotes.Search(movie ?? string.Empty, actor ?? string.Empty);       
    }
}